fix: corrigir dados de pro-rata em downgrade e renovação de contratos
- Corrigir PaymentService para calcular prorated_old corretamente em contratos expirados
- Implementar lógica de downgrade conforme especificação TEST_MVP.md
- Garantir que campos monetários sejam salvos em reais (não centavos)
- Corrigir testes para validar cenários de downgrade corretamente
- Remover criação automática de pagamentos na criação de contratos
- Todos os 27 testes passando (100% sucesso)

Problemas resolvidos:
- Pagamentos sem dados de pro-rata no histórico
- Cálculos incorretos em cenários de downgrade
- Valores zerados nos campos prorated_old e prorated_new
- Testes com expectativas erradas de autenticação

+++ api/app/DTOs/PaymentDTO.php
@@ -15,9 +15,9 @@ class PaymentDTO
         public Money $amount,
         public Carbon $payment_date,
         public PaymentStatus $status,
-        public ?int $discount_applied = null,
-        public ?int $prorated_old = null,
-        public ?int $prorated_new = null,
-        public ?int $applied_credits = null,
+        public ?float $discount_applied = null,
+        public ?float $prorated_old = null,
+        public ?float $prorated_new = null,
+        public ?float $applied_credits = null,
     ) {}
 }
